+++ docs/changelog.rst
@@ -18,6 +18,11 @@ Each revision is versioned by the date of the revision.
 - Add documentation on how to upgrade the MediaWiki charm.
 - Add documentation on how to rotate secrets managed by the MediaWiki charm.
 
+2026-05-20
+----------
+
+- Implement SAML authentication using the SimpleSAMLphp extension via the saml interface.
+
 2026-05-07
 ----------